requirements.txt
pypdf
langchain
chromadb # Vector storage
pytest
boto3
langchain_ollama
langchain_chroma
customtkinter
pillow
pywinstyles